+++ src/main/scala/com/github/frankivo/CovidBot.scala
@@ -8,7 +8,8 @@ object CovidBot {
     val akka = ActorSystem()
 
     val stats = akka.actorOf(Props(new CovidStats))
-    akka.actorOf(Props(new Telegram(stats)))
+    val updater = akka.actorOf(Props(new CovidStats))
+    akka.actorOf(Props(new Telegram(stats, updater)))
   }
 
 }
